requirements.txt
#python=3.6

opencv-python
imagehash
simplekml
flickrapi
exifread
pyopencl
gpsphoto
requests
piexif
pillow
geopy
tqdm
pocl
silx
scipy==1.1.0